LPS-1736
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@25351 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/velocity/VelocityEngineImpl.java
@@ -55,7 +55,9 @@ public class VelocityEngineImpl implements VelocityEngine {
 		StringResourceRepository stringResourceRepository =
 			StringResourceLoader.getRepository();
 
-		stringResourceRepository.removeStringResource(resource);
+		if (stringResourceRepository != null) {
+			stringResourceRepository.removeStringResource(resource);
+		}
 	}
 
 	public VelocityContext getEmptyContext() {
